+++ packages/events/src/router/middleware/validate/index.ts
@@ -29,17 +29,70 @@ import {
   deepMerge,
   deepDropNulls,
   omitHiddenFields,
-  EventState
+  EventState,
+  Inferred,
+  isFieldVisible,
+  errorMessages,
+  runFieldValidations
 } from '@opencrvs/commons/events'
 import { getEventConfigurationById } from '@events/service/config/config'
 import { getEventById } from '@events/service/events/events'
 import { ActionMiddlewareOptions } from '@events/router/middleware/utils'
 import {
-  getFormFieldErrors,
   getInvalidUpdateKeys,
   getVerificationPageErrors,
   throwWhenNotEmpty
 } from './utils'
+
+export function getFieldErrors(
+  fields: Inferred[],
+  data: ActionUpdate,
+  declaration: EventState = {}
+) {
+  const visibleFields = fields.filter((field) =>
+    isFieldVisible(field, { ...data, ...declaration })
+  )
+
+  const visibleFieldIds = visibleFields.map((field) => field.id)
+
+  const hiddenFieldIds = fields
+    .filter(
+      (field) =>
+        // If field is not visible and not in the visible fields list, it is a hidden field
+        // We need to check against the visible fields list because there might be fields with same ids, one of which is visible and others are hidden
+        !isFieldVisible(field, data) && !visibleFieldIds.includes(field.id)
+    )
+    .map((field) => field.id)
+
+  // Add errors if there are any hidden fields sent in the payloa
+  const hiddenFieldErrors = hiddenFieldIds.flatMap((fieldId) => {
+    if (data[fieldId as keyof typeof data]) {
+      return {
+        message: errorMessages.hiddenField.defaultMessage,
+        id: fieldId,
+        value: data[fieldId as keyof typeof data]
+      }
+    }
+
+    return []
+  })
+
+  // For visible fields, run the field validations as configured
+  const visibleFieldErrors = visibleFields.flatMap((field) => {
+    const fieldErrors = runFieldValidations({
+      field,
+      values: data
+    })
+
+    return fieldErrors.errors.map((error) => ({
+      message: error.message.defaultMessage,
+      id: field.id,
+      value: data[field.id as keyof typeof data]
+    }))
+  })
+
+  return [...hiddenFieldErrors, ...visibleFieldErrors]
+}
 
 function validateDeclarationUpdateAction({
   eventConfig,
@@ -88,7 +141,7 @@ function validateDeclarationUpdateAction({
     .filter((page) => isPageVisible(page, cleanedDeclaration))
     .flatMap((page) => page.fields)
 
-  const declarationErrors = getFormFieldErrors(
+  const declarationErrors = getFieldErrors(
     allVisiblePageFields,
     cleanedDeclaration
   )
@@ -105,10 +158,7 @@ function validateDeclarationUpdateAction({
     deepDropNulls(annotation ?? {})
   )
 
-  const annotationErrors = getFormFieldErrors(
-    reviewFields,
-    visibleAnnotationFields
-  )
+  const annotationErrors = getFieldErrors(reviewFields, visibleAnnotationFields)
 
   return [...declarationErrors, ...annotationErrors]
 }
@@ -136,7 +186,7 @@ function validateActionAnnotation({
   )
 
   const errors = [
-    ...getFormFieldErrors(formFields, annotation, declaration),
+    ...getFieldErrors(formFields, annotation, declaration),
     ...getVerificationPageErrors(visibleVerificationPageIds, annotation)
   ]
