+++ packages/events/src/router/middleware/validate/utils.ts
@@ -11,69 +11,12 @@
 
 import { TRPCError } from '@trpc/server'
 import _ from 'lodash'
-import {
-  ActionUpdate,
-  Inferred,
-  errorMessages,
-  runFieldValidations,
-  isFieldVisible,
-  EventState
-} from '@opencrvs/commons/events'
+import { ActionUpdate } from '@opencrvs/commons/events'
 
 type ValidationError = {
   message: string
   id: string
   value: unknown
-}
-
-export function getFormFieldErrors(
-  formFields: Inferred[],
-  data: ActionUpdate,
-  declaration: EventState = {}
-) {
-  const visibleFields = formFields.filter((field) =>
-    isFieldVisible(field, { ...data, ...declaration })
-  )
-
-  const visibleFieldIds = visibleFields.map((field) => field.id)
-
-  const hiddenFieldIds = formFields
-    .filter(
-      (field) =>
-        // If field is not visible and not in the visible fields list, it is a hidden field
-        // We need to check against the visible fields list because there might be fields with same ids, one of which is visible and others are hidden
-        !isFieldVisible(field, data) && !visibleFieldIds.includes(field.id)
-    )
-    .map((field) => field.id)
-
-  // Add errors if there are any hidden fields sent in the payloa
-  const hiddenFieldErrors = hiddenFieldIds.flatMap((fieldId) => {
-    if (data[fieldId as keyof typeof data]) {
-      return {
-        message: errorMessages.hiddenField.defaultMessage,
-        id: fieldId,
-        value: data[fieldId as keyof typeof data]
-      }
-    }
-
-    return []
-  })
-
-  // For visible fields, run the field validations as configured
-  const visibleFieldErrors = visibleFields.flatMap((field) => {
-    const fieldErrors = runFieldValidations({
-      field,
-      values: data
-    })
-
-    return fieldErrors.errors.map((error) => ({
-      message: error.message.defaultMessage,
-      id: field.id,
-      value: data[field.id as keyof typeof data]
-    }))
-  })
-
-  return [...hiddenFieldErrors, ...visibleFieldErrors]
 }
 
 export function getVerificationPageErrors(
